+++ desmume/src/windows/File_Extractor/unrar/coder.cpp
@@ -18,9 +18,10 @@ void RangeCoder::InitDecoder(Unpack *UnpackRead)
 }
 
 
+// (int) cast before "low" added only to suppress compiler warnings.
 #define ARI_DEC_NORMALIZE(code,low,range,read)                           \
 {                                                                        \
-	while ((low^(low+range))<TOP || range<BOT && ((range=-low&(BOT-1)),1)) \
+  while ((low^(low+range))<TOP || range<BOT && ((range=-(int)low&(BOT-1)),1)) \
   {                                                                      \
     code=(code << 8) | read->GetChar();                                  \
     range <<= 8;                                                         \
